* Headers/AppKit/NSColorPanel.h: Change -alpha to CGFloat. 	* Source/NSColorPanel.m: Remove left over GModel loading code. 	Set default mode to colour wheel. 	Small cleanups mostly in comments.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@40229 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSColorPanel.h
@@ -132,7 +132,7 @@ enum {
 	 fromView:(NSView *)sourceView;
 - (void)setColor:(NSColor *)aColor;
 
-- (float)alpha;
+- (CGFloat)alpha;
 - (NSColor *)color;
 
 @end
